Update patch set 5
Patch Set 5:

> > Also see my comment on Jan 24 7:02 PM.
 > 
 > FYI Gerrit shows localized times, so I don't see a "Jan 24 7:02 PM"
 > fro you ;) but there's only 1 feedback from you yesterday on
 > January 24 with 8 comments, which I've all tried to address, or did
 > I forget any of the points?

This one:

This patch intends to provide a uniform naming format which is good. Based on your comments in other patches, I got the impression this patch is also intended to avoid users doing: 

  Map<String, Counter>  counterMap;

If that is correct then the proposed Labeled interface always creates a new instance and does not store instances so users will still need to maintain a Map. This is OK but I'm wondering if we could also provide that functionality here, since it appears to be a common pattern,  with a LabeledMetricCache (or Map or something else - I know you don't like term "Cache" here :))

Patch-set: 5