Update patch set 5
Patch Set 5:

> This one:

oups, I genuinely missed that one, for some reason; sorry.

 > If that is correct then the proposed Labeled interface always
 > creates a new instance and does not store instances so users will
 > still need to maintain a Map. This is OK but I'm wondering if we
 > could also provide that functionality here, since it appears to be

oh... wait, yeah that's actually an *EXCELLENT* point; thank you!  I indeed did mean for it to be used like that, see the use of meterWithOneDynamicLabel and meterWithTwoDynamicLabels in the MetricsExample... and you are pointing out that as is that is actually broken; I'll fix that, and cover it with real TDD - thanks again for noting!!

> How about passing name/value pairs then, ie
> T label(String labelName, String labelValue);
> This way label names and values are coupled together rather than
> passing the names to newMeter and the values to cascaded label calls.

right, but an API like this would "defer" the creation of labels until what to me seems too late, at usage time when you have values for the label.  With how we have it now, see MetricsExample, you create a Metric and specify the label/s (names) at creation time, and then give those labels values at usage during later run-time - and we can still statically enforce that intended number of labels given names to match the values provided.  So the system knows at boot time what metrics with exactly what labels are available.

Patch-set: 5