Update patch set 6
Patch Set 6:

> But you don't store the MetricDescriptor and label names or is the
 > intention to do so later? Also what would do you intend to use that
 > for, reporting?

That's true.. ;) yeah I was hoping to build a "metrics:list" kind of CLI command one day, which would print these labeled metrics in groups.

Skitt also had an interesting idea that in say MetricsExample one subclasses Labeled to be able to write like meterWithOneDynamicLabel.port("456") with the API as it is now.

Patch-set: 6
Label: Verified=0