JobCoordinatorImpl temporarily remains backward compatible
This amends the previous (re-re-re-reverted) change to not break genius
and netvirt, so that we can merge this into infrautils now, and then get
the genius I80e3f5b6e8a601345cd2fbae79f0de22d0fa0e7d and netvirt
If3c3bcdb8a972c7e57cff20fd24d5addefa7746b changes in separately whenever
is convenient.

This approach additionally and perhaps even more importantly also has
the advantage that pending in-flight changes in genius and netvirt will
not be broken and require rebases, as in the original plan.

The idea is to revert this commit again once we're all through; maybe
next week, or maybe even leave it in for Oxygen and only revert this at
the very start of Fluorine.

Change-Id: I3a7cf3cccca9488f85468e675dd2f31f7f167c33
Signed-off-by: Michael Vorburger <vorburger@redhat.com>