Update patch set 4
Patch Set 4:

Thanks.  I'm adding 2 other infrautils committers so that we can wrap this up and merge ASAP (we should definitely get this still into Oxygen IMHO).  It's basically OK for me as it is now, but there's 1 point I feel we should clarify... what do we want the out-of-the-box default for this fileReporterIntervalSecs to be?  You currently have it set to 120 (in class Configuration), which is what the current master code already has hard-coded.  So we will write files into the data/metrics directory every 2 minutes.  I'm not sure if there is any clean up of old files, to avoid disks filling up?  Either way, is that what we want as the out-of-the-box default?  Or do we want this to be disabled by default after installation?  With people being able to enable it as needed by editing etc/org.opendaylight.infrautils.metrics.cfg?  I'm myself not sure what the right default should (but would lean towards disabled by default, but could be convinced otherwise), so thought worth briefly discussing and getting agreement on, before merging this... what are your and others' reviewers views on this?  Also, whatever we decide, I'll make the value in the class Configuration and the (commented out) value in org.opendaylight.infrautils.metrics.cfg the same.

Patch-set: 4
Label: Verified=0