Update patch set 4
Patch Set 4:

Do we want to use the cfg files or the blueprint xml files? NetVirt and Genius had gone down the path of the xml files and openflowplugin and ovsdb went with cfg's. The two work similar but there are fundamental differences. My understanding was that the xml files were more useful to have and the only benefit to the cfg's was that it was a text file but you had to duplicate the file on each ODL node.

Patch-set: 4