use feature for inter-project dependency
Use feature to express inter-project dependency correctly instead of
bundle.

For details, refer to
https://lists.opendaylight.org/pipermail/release/2018-January/013604.html
https://lists.opendaylight.org/pipermail/release/2017-September/012356.html

More specifically, the use of <scope>provided for the bundles-test-lib
which is an <Embed-Dependency> avoids the automatic feature generation
to have a wrap for bundles-test-lib (from odlparent), which actually
never needed in the the first place (that wrapped JAR was never used),
see https://jira.opendaylight.org/browse/INFRAUTILS-28 for full analysis.

Change-Id: Ibacbe16b58ec5f631867284e1ef6708a95a590ec
Signed-off-by: Faseela K <faseela.k@ericsson.com>