add new API for programmatic registration of web Servlet, Filter, etc.
implementation(s) for this will be contributed in follow-up changes

The purpose of this API is to let projects with web components, such as
neutron, aaa, restconf, ditch their respective web.xml.  This will have
a number of advantages, some of which are documented in the JavaDoc of
the new WebServer interface and WebContext class.

Change-Id: Ib2df87ca31a2bde547efbf73e0475a1cd64ea6ea
Signed-off-by: Michael Vorburger <vorburger@redhat.com>