make ThreadsWatcher log at fixed intervals, not continuously
yet let it still check at shorter intervals than it will log

because it's been observed that when we hit max threads or deadlock,
then as it is, it will log continuously like crazy, at the speed of the
intervals it checks (defaults to 500ms).  Increasing that interval to a
much larger value (say 1min) is a shame because that could then take a
longer time to find issues, or even miss them.  So that this does it can
keep checking frequently, but doesn't have to report and spew logs as
often.

The new max log intervals are hard-coded to 1min in this first change.

see https://lists.opendaylight.org/pipermail/infrautils-dev/2018-February/000596.html

Change-Id: I075d9789af7c091e67b75c390963e71351f60672
Signed-off-by: Michael Vorburger <vorburger@redhat.com>
(cherry picked from commit ad048e9d07fa4154f2363a43b99c10a46f85e5de)