INFRAUTILS-16
JobCoordinator enqueueJob must return a ListenableFuture instead of
void.

With this patch, the enqueueJob in JobCoordinator would return a list of
futures which the clients(which had enqeued the jobs) could use to know
the success/failure of each job and act on it(instead of only retrying
the failed jobs).

Change-Id: I96304fdd386a82d4f8e3648cfd3293b2ab3348bc
Signed-off-by: gobinath <gobinath@ericsson.com>