Update patch set 1
Patch Set 1:

well, I tried to make this work in OFP which is kind of what started it.
you can see that patch here:

  https://git.opendaylight.org/gerrit/c/70221/

the OFP bug:

  https://jira.opendaylight.org/browse/OPNFLWPLUG-988

It sort of does what we would expect, except for the output is weird to
me. Not sure where to dig in to.

without the 70221 patch, this new errorCause stuff looks like this:

    karaf@root()> showsvcstatus
    Timestamp: Fri Mar 30 21:13:54 PDT 2018
    Node IP Address: 127.0.0.1
    System ready state: ACTIVE
      OPENFLOW            : ERROR             Optional.empty
    karaf@root()>

Obviously that Optional.empty is kind of dumb to see.

with the OFP partner change to take this errorCause in to account, it
looks like this:

    karaf@root()> showsvcstatus 
    Timestamp: Fri Mar 30 21:08:01 PDT 2018
    Node IP Address: 127.0.0.1
    System ready state: ACTIVE
      OPENFLOW            : ERROR             Optional[java.io.IOException: some switch connections failed to start]
    karaf@root()>


The details we want is present ("some switch connections failed to start"), but
why is it in brackets with Optional[]?

Patch-set: 1