Update patch set 1
Patch Set 1:

[adding infrautils committer Tom as an additional reviewer]

 > It sort of does what we would expect, except for the output is
 > weird to me. Not sure where to dig in to.

so that is coming from this which we together added (here) in DiagStatusServiceMBeanImpl acquireServiceStatusDetailed:

    .append(status.getErrorCause())

 > OPENFLOW            : ERROR             Optional.empty
 > Obviously that Optional.empty is kind of dumb to see.

so we probably want to only do that append doing something like this?

    if(status.getErrorCause().isPresent()) {
        statusSummary.append(status.getErrorCause().get());
    }

but first read on next for an even nicer way...

 > OPENFLOW            : ERROR             Optional[java.io.IOException:
 > 
 > The details we want is present ("some switch connections failed to
 > start"), but why is it in brackets with Optional[]?

you can use the get() on the Optional if (and only if) it isPresent(), as shown above. Better yet, you can use a "lamba function" and instead use a syntax going something like this:

    status.getErrorCause().ifPresent(cause ->
        statusSummary.append(cause));

but if we have a Throwable cause we really should also show its full getStackTrace() via printStackTrace() ... because Throwable.toString() only prints its getMessage(), so we would loose 3/4 of the additional information which we now have available here, and that's a shame.

Patch-set: 1