Update patch set 6
Patch Set 6:

(4 comments)

> 1. In case of no retries required(retry=0), do we still need to
 > have the JobCallBack since we wont retry the task anyway? This

probably OK, but I'm not 100% sure; but is this required for / part of this change, or anyway a bit separate from this change here? How about you raise a separate new review in parallel making this change, and add the same reviewers as are on this change, and we can look further at this point there?

Patch-set: 6
Label: Verified=0