Update patch set 5
Patch Set 5:

> (1 comment)
 > 
 > > > Will this be put into odlparent at some point?
 > >
 > > I don’t think ODL Parent is the right place for this...
 > 
 > Why not? odlparent has the other checkstyle/findbugs configs etc.

Right, it has *config*, but no implementation of checks. The way I see it, this check is intended to be used to track code migrations; in the case I’m pursuing, I want to prevent new uses of DataBroker::newReadWriteTransaction and DataBroker::newWriteOnlyTransaction. That involves very specific Checkstyle configurations which wouldn’t belong in ODL Parent anyway; I’d be adding the *config* for this in MD-SAL or in the specific projects; so there’s no need to have this in ODL Parent.

Patch-set: 5
Label: Verified=0