let MBeanUtils.invokeMBeanFunction throw problems, not swallow and log
and handle the impact of this in the DiagStatusCommand.  Normally Karaf
will correctly inform the CLI user and log any exceptions which command
Action implementations throw; that's why execute() throws Exception. But
in this specific case, the DiagStatusCommand even already has a catch
(Exception e) to handle JMX Operation failures - so all the more reason
to let it bubble up to there.

Change-Id: I29af95d5eaf204718f0d8eb95c45c6bde4fe87c2
Signed-off-by: Michael Vorburger <vorburger@redhat.com>