let MBeanUtils' readMBeanAttribute throw problems, not swallow and log
actually we cannot do this immediately, because it's used in
DatastoreServiceStatusProvider in genius; so instead let's @Deprecated
readMBeanAttribute() and introduce getMBeanAttribute() throws
JMException.

we also change ClusterMemberInfoProvider to use getMBeanAttribute()
instead of readMBeanAttribute(), and handle any JMException here in the
caller, instead of in the utility.  This ultimately changes nothing from
an end-user PoV, but the utility class has a better API like this.

in a next step, we can use the new method in
DatastoreServiceStatusProvider in genius, and then remove the
@Deprecated.

Change-Id: Iaf5edf99ad21653a939cc3c5d9c099c71d4f2676
Signed-off-by: Michael Vorburger <vorburger@redhat.com>