Update patch set 2
Patch Set 2: Code-Review-1

> Build Unstable
 > 
 > https://jenkins.opendaylight.org/releng/job/infrautils-patch-test-genius-fluorine/1/
 > : UNSTABLE (skipped)
 > 
 > No problems were identified. If you know why this problem occurred,
 > please add a suitable Cause for it. ( https://jenkins.opendaylight.org/releng/job/genius-csit-1node-gate-all-fluorine/46/
 > )
 > 
 > Logs: https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/infrautils-patch-test-genius-fluorine/1

so yeah, this patch fails genius csit because it's based on my patch
that introduces the .get() which gives us the problem with showsvcstatus.

https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/genius-csit-1node-gate-all-fluorine/46/robot-plugin/log.html.gz#s1-s1-k1-k1-k1-k1-k1-k1-k1-k1-k1-k6

if we want this patch by itself, we can just rebase it on top of
master, then double check with csit if you like.

Patch-set: 2
Label: Code-Review=-1