Update patch set 2
Patch Set 2:

> > ok - but does TestableScheduledExecutorService really need all of
 > > this and do we really need TestableScheduledExecutorService (TBH,
 > I
 > > found it just as easy to use Mockito directly plus I wanted to
 > > verify the ScheduledFuture was closed)?
 > 
 > I find this approach (Partials) very handy.  FYI we're already
 > using this in a number of other tests to create "test fakes" (e.g.
 > genius' TestIMdsalApiManager, and some others) - it just saves
 > having to implement lots of empty methods.

yes it's useful but TestIMdsalApiManager et al are using the yangtools utilities. So it would be only infrautils that would need the copied yangtools utilities. I think you'd agree copying isn't ideal and should only be done if absolutely needed and compelling. To me the TestableScheduledExecutorService doesn't fit that criteria.

Patch-set: 2