add guice (and mycila) versions explicitly
so that we can bump them here, instead of having to
wait 1 year for a new odlparent version.  It will be
removed from odlparent in c/72222, so that project inherit it from
here (no project should have a direct dependency to Guice).

Guice 4.2.0 update is out (we're on 4.1.0), but it's
a *huge* PITA to use it - because odlparent dependencyManagement
trumps transitively inherited dependencies... :-(

Change-Id: I2c2f6df7b19d8b1662da6a6fad88c8caca2945a0
Signed-off-by: Michael Vorburger <vorburger@redhat.com>