Update patch set 5
Patch Set 5:

> 1 CSIT failure is not due to this patch, and it is infra related
 > ERROR.

agreed that the CSIT failure is most likely nothing to do with this patch, but I am
not sure this is INFRA related. Is there anything tracking this prolem (JIRA)? The top
level problem is a timeout trying to communicate with ODL via REST. If this was
truly and infra problem, I'm sure that kind of thing would be happening all over
in other CSIT too. I haven't heard of that. It's not happening in netvirt csit, afaict.

Look at the karaf.log. Seems like some troubles are there:

    org.opendaylight.genius.idmanager.IdManagerException: Ids exhausted for pool : test-pool


    org.opendaylight.mdsal.common.api.OptimisticLockFailedException: Optimistic lock failed for path /(urn:opendaylight:genius:itm:config?revision=2016-04-06)tunnel-monitor-params/enabled


ignore the NPEs around the karaf ssh util. that is a karaf bug with a fix we'll get on
the next karaf upgrade.

anyway, I would not be ignoring these CSIT issues as infra related without making
the releng/integration team aware of them, if that's really what we think they are.

From my perspective, infra has been pretty solid for a while now after all the
kinks were ironed out when we moved cloud providers.

JamO

Patch-set: 5
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0