Update patch set 1
Patch Set 1:

Abandoning this during my big clean up before the lazy european ;) summer, and more importantly to reduce my own “mental load”... given that I've not managed to wrap this one up in so long, clearly it's not really all that critical in the bigger picture of things. It could always be Restored, by me - or anyone else interested in picking this up. Onwards. Focused.

Patch-set: 1