make diagstatus ServiceDescriptor description optional
because if you are reporting e.g. ServiceState STARTING or OPERATIONAL
then it's dumb to be forced to have to provide an "artificial"
description - what is one going to put; something like "Service is
starting" is just a useless duplication of the ServiceState.

The description is still useful if you actually have additional
information to provide, I guess typically perhaps e.g. in case of a
SUSPECTED ServiceState.

Change-Id: I22109f46e4505ec4336f5d8fdc1794bbb8d05562
Signed-off-by: Michael Vorburger <vorburger@redhat.com>