refactor MBeanUtils' invokeRemoteJMXOperation()
It used to be a diagstatus specific method hard-coded to
DiagStatusServiceMBean acquireServiceStatusDetailed(), and now is a
general utility.

This is one of many more steps required to enact the TODO in MBeanUtils
to, potentially, if there is any need for it for anything else (and I am
not at all convinced that there actually is), move MBeanUtils out of
diagstatus into general utils.

Change-Id: I7d339c66335994c3baa9d110a36d980269119cff
Signed-off-by: Michael Vorburger <vorburger@redhat.com>