make Executers return Listening[Scheduled]ExecutorService
instead of plain [Scheduled]ExecutorService.  This is very useful
(required) to make it obvious to use ListenableFutures's addErrorLogging
for the futures returned by the submit methods.

we, unfortunately, cannot change the return type in one go, even though
Guava's Listening* variants extend the original ones, because it breaks
compilation in other projects without a rebuild/remerge.  Therefore we
have to @Deprecate the old ones and introduce new ones.

Change-Id: I6bcb9e969ae446a83e10b6225c652bdc5271eae5
Signed-off-by: Michael Vorburger <vorburger@redhat.com>