Update patch set 1
Patch Set 1:

(1 comment)

> (1 comment)
 > 
 > I'm not sure of the usefulness of a class with a single method that
 > is immediately deprecated that merely wraps Futures.makeChecked.
 > Why not just inline it or remove the use of CheckedFuture since
 > it's deprecated?

see https://lists.opendaylight.org/pipermail/genius-dev/2018-August/003220.html for background :

it's useful to transition existing methods which return CheckedFuture to FluentFuture because if those are APIs, like in that email thread, we cannot "just" remove them in one go.

I do believe that this has value for "discoverability", it seems trivial when looking at this change, but as per thread, quote : "at least for me it's often still confusing what is what among all those different Future kind of types - so I like have easy to find conversion methods. This is consistent with
other such methods which are available in infrautils.utils.concurrent; the
pattern is always that you simply look for a class ending in *s for the
type you have, and will find a toXYZ() method on it for the type you want."

Edwin, Faseela please chime in here with votes & comments if this has value.

Patch-set: 1
Label: Verified=0