Update patch set 1
Patch Set 1:

+skitt, because he looked at something related to this.

Assuming that this isn't a Blocker (doubtful), I'm reluctant to accept this change for Fluorine mere hours before the branch is cut, just out of general carefullness, and want to keep this open to discuss getting it into Neon.

Is it just "nice" not to have Immutables at runtime, or does it.. "fix" anything?  Just trying to understand why you are doing this, if I may ask.  (I'm not against it, just trying to understand.)

Also, and for Stephen to chime in on (when back from PTO), IMHO it would be much better if we could somehow solve this in odlparent... we'll add Immutables dependencies elsewhere, and forget to do the exclusion.. if for some reason it's impossible to solve this in odlparent, then I'm OK to merge this is a pragmatic local fix - but if it could just be transparently taken care of in odlparent, that would be nicer.

Patch-set: 1
Label: Code-Review=0