Update patch set 5
Patch Set 5:

(2 comments)

I like the direction this is taking, but it seems to be only partial - there is still a pax-cdi-api in caches' sample/impl and test/ (and an @OsgiServiceProvider and @OsgiService in SampleServiceWithCachingImpl) - is that intentional?  Seeing this also gave me another idea perhaps worth discussing in this context: With what we are up to here now, we really should be removing pax-cdi-api from odlparent 4.0.0 all together?

Patch-set: 5