ditching ClusterMemberInfoProvider getSelfAddress() Optional
That seems weird - we should always be able to know our own node
address.  The use of this in the current implementation is totally
inconsistent anyway - note how DiagStatusServiceMBeanImpl would just not
startRMIConnectorServer() if there were no selfAddress; but in
DiagStatusCommand we would do .orElse("localhost")...

It seems simpler to "fail fast" and clearly during boot if we cannot
even obtain our own self address.  The known problem with this not
working on IPv6 is just getting fixed in parallel anyway.

This is in preparation for making ClusterMemberInfoProvider non-static,
which is in preparation for an upcoming DiagStatusCommandTest (motivated
by the IPv6 fix in this code).

Change-Id: I121b2035aa0e66a815fc03547301c23f643f1f20
Signed-off-by: Michael Vorburger <vorburger@redhat.com>