factor out new SystemReadyImpl from KarafSystemReadyImpl
so that a non-OSGi ODL distribution can provide SystemReadyMonitor (it
would call ready() after whatever it uses for dependency injection is
fully done).

This comes out of https://github.com/vorburger/opendaylight-simple/,
which currently has a SystemReadyBaseImpl class that is a poor man's
version of this SystemReadyImpl; we will replace that with this.

This new SystemReadyImpl also is (replaces) what I made in for that TODO
commented out EXPLICIT Behaviour in the TestSystemReadyMonitor, which
therefore gets removed with this.

Change-Id: Icff397639f0c245b98e8e3b32dee4b3360ceaa6e
Signed-off-by: Michael Vorburger <vorburger@redhat.com>