simplify Exception handling in SimpleSystemReadyMonitor
and strictly speaking changing it's behaviour slightly; it used to
re-throw RuntimeException but not Exception, but not does not anymore.

This distinction made sense while this code was in what is now
KarafSystemReadyImpl, where testBundleDiag.checkBundleDiagInfos may
throw a SystemStateFailureException but never a RuntimeException, but in
the SimpleSystemReadyMonitor it's better to keep this simpler; caller of
ready(), like Dependency Injection framework boot sequence of
opendaylight-simple, do not expect to receive RuntimeException at this
stage, from an SDN application's registered ready listeners.

Change-Id: I2ad9d2a691d2f1d88c4f574991f4bb61f5adb90a
Signed-off-by: Michael Vorburger <vorburger@redhat.com>