Update patch set 1
Patch Set 1:

> Michael, this is not solving the problem seen in genius CSIT as
 > Robert said..could you please check the karaf logs to see if the
 > TCCL is retained or not?

The code looks right - it seems this is just not fixing the issue b/c for the caller that fails, when the job is enqueued, the TCCL is probably not getting set anywhere or is not a ClassLoader that can "see" the augmentation class. Even so it seems this patch is still reasonable.

Patch-set: 1