Update patch set 5
Patch Set 5:

(1 comment)

> "Component must have a valid id" occurs due to the <command-bundle>
 > - its NamespaceHandler tries to extract the scope from the @Command
 > annotation in DiagStatusCommand but doesn't find it. This is b/c
 > DiagStatusCommand is using the new(er) org.apache.karaf.shell.api.action.Command
 > but the NamespaceHandler is looking for the deprecated
 > org.apache.karaf.shell.commands.Command. So you can't use
 > <command-bundle> unless DiagStatusCommand is downgraded to use the
 > deprecated org.apache.karaf.shell.commands.Command et al API.
 > 
 > Looks like you added <command-bundle> in order to inject 2
 > additional fields  however apparently this needs to be done via
 > @Reference with the new API. The properties Map may be problematic
 > to inject via @Reference - if so I'd suggest to wrap it in a class.
 > The other option is to use <command-bundle> and downgrade
 > DiagStatusCommand to the deprecated API.

Thanks a lot Tom! This issue is fixed now, but am still having issues with loading of pax web properties. it is coming as empty. The way am using persistent-id is correct in blueprint.xml? Or is there anything additionally I have to do to get this working? As of now I just added in blueprint.xml, as I could understand from documentations available.

Patch-set: 5
Label: Verified=0