Update patch set 1
Patch Set 1:

> I think callers should use GeneralEntity.getIdentifier() instead.

but that returns an InstanceIdentifier, not a String.  You probably mean IFF JobCoordinator key was an InstanceIdentifier instead of String, then callers could use entity.getIdentifier() instead of entity.toString() ?

 > ... if it's even possible. It seems the hierarchical "Utils" design
 > across projects leaves quite a bit to be desired.

Huh?

 > Michael, why can’t infrautils depend on yangtools?

The old discussion that infrautils should not depend on yangtools, but perhaps (in the future, if infrautils is MRI; or yangtools no longer MRI) then if anything yangtools should depend on infrautils.

The real motivation for INFRAUTILS-60 was GENIUS-237, which can be resolved simpler; I'm therefore abandoning this for now and adressing GENIUS-237 differently; we can resurrect this in the future, if there anyone has more appetite to resolve this.

Patch-set: 1
Reviewer: Gerrit User 5867 <5867@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0