deprecate getAllServiceDescriptorsAsJSON() and offer to/from/Json()
replace all usages of this deprecated method in the diagstatus code
itself already, but do not yet remove it - just in case there are other
users.

This is a better OOP design as it provides better encapsulation.

More importantly, it allows to fix the concurrency bug in
DiagStatusServlet by creating the JSON from the ServiceStatusSummary.

Change-Id: I7f5601912098e5fe0674a39849a36ac5105e33a0
Signed-off-by: Michael Vorburger <vorburger@redhat.com>