Add errorprone for Java 9+
errorprone is now recommended as a javac plugin, which works
just fine with JDK9+.

Using that approach for JDK8 has two problems:
- it requires uglification of arguments, as JDK8 does not grok
  multiline arguments properly
- it requires forking the compiler and thus breaks other
  annotations processors

We therefore retain the original JDK8 profile and add a JDK9/10
profile, which uses the javac approach.

The new profile will work with JDK11, too, except it requires
errorprone 2.3.2, but we cannot do that, as that release has
a bug, which we trigger.

Change-Id: Ic954a55dc8fba99426e2f9dfec0b70a299cb69ab
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>