Update patch set 7
Patch Set 7:

The recommended way is in revision 3 of this patch: https://git.opendaylight.org/gerrit/#/c/79211/3

As you can see there is general weirdness in layout (due to Java 8), which would be fine if it worked, but it does not.

The reason for that is that annotationProcessorPaths directly interferes with drop-in annotation processors (like immutables.org), which no longer get run, leading to compilation failures.

I am pretty sure annotationProcessorPaths is needed for -Xplugin to work on JDK <= 10, hence we cannot ditch it. Perhaps we can for JDK11, which would restore sanity -- but I cannot confirm that.

So this is not ready to go in at this stage.

Patch-set: 7