Update patch set 8
Patch Set 8:

> Patch Set 7:
> 
> > Patch Set 3: Code-Review-2
> > 
> > Ugh, this is wrong:
> > - Apache license, our build does the wrong thing by default
> > - it should be clearly noted this is third-party code, such as a subdirectory
> > - it is clearly a self-contained project, hence it should not be just dumped into infrautils
> 
> Robert, pls take a look at the latest patch and let me know some specific changes that i can make to get this merged. thanks.

Well, I do not want to have this merged. It is the wrong approach -- how about lumina forking micylla on github and issue a release in central? We are consuming things like tech.pantheon.triemap and ch.vorburger.xtendbeans that way.

It has the benefit of serving a wider community (not just ODL) and retaining the commit history, and not growing infrautils by 50% of its current size.

Patch-set: 8