Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > Well, I do not want to have this merged. It is the wrong approach -- how about lumina forking micylla on github and issue a release in central? We are consuming things like tech.pantheon.triemap and ch.vorburger.xtendbeans that way.
> > 
> > It has the benefit of serving a wider community (not just ODL) and retaining the commit history, and not growing infrautils by 50% of its current size.
> 
> Thanks for clarifying Robert. In principle, making components of ODL projects depend on such third party GitHub projects has a side effect that the community cannot exercise any influence on those, being external. Isn't that a risk? It would be better to contain in odl any such code including the examples you cited above.

This is a question of maintenance. ODL is a siloed project from the perspective of the world -- i.e. you cannot easily collaborate on ODL projects via GitHub.

At the end of the day, we are just consuming artifacts from Central -- we do not care what governance they go through, really, as long as they work for us. We choose on what to depend on a particular project's health -- and we did switch from dead upstreams couple of times.

Exercising influence in the FOSS world means putting boots on the ground -- that way you ahve influence without having to fork and play other shenanigans.

Finally, ODL has a mission, and that mission is not maintaining Guice extensions, but rather something a quite a bit different. We should strive to minimize the amount of code we need to maintain -- otherwise it will just choke us.

Having said that, if you feel that this code needs to be forked under ODL umbrella, feel free to propose a self-contained project (guice-di?), make the commitment to maintain it when the project is approved, please do take all the guice things from infrautils there, as it would be a natural (and well-scoped!) place for them to live.

Patch-set: 8