Update patch set 8
Patch Set 8:

I noticed that you seem to have changed the license in the Mycila Java files from their original ASL header to EPL.  You can keep the original ASL license headers everywhere, and exclude the Checkstyle check; there's precedent for doing that elsewhere in ODL, e.g. check this out:

https://github.com/opendaylight/odlparent/blob/master/karaf-plugin/pom.xml#L84

https://github.com/opendaylight/odlparent/blob/master/karaf-plugin/src/main/java/org/opendaylight/odlparent/PopulateLocalRepoMojo.java

While I agree with Robert's point in a general open source sense, in this particular case, for these ca. 20 utility classes, I personally don't see what the big deal is to "inline fork" them into ODL... my understanding is that this is both legally perfectly OK, as well as not the first time in the history of open source that one project copies code from another defunct one.  It's not like there are millions of people volunteering to maintain Mycila in another repo.  To me, it thus seems better for future maintainability to just have it in ODL than having someone like Lumina maintain it in an external repo on GitHub, and go through the trouble of having to release it on Maven central; future OpenDaylight maintainers then will have the exact same problem with as with github.com/mycila/guice - so it's GOOD to have this here.  And creating yet another new ODL git repo & sub-project for putting these 20 classes into it seems total overkill from a process point of view to me.

I'll therefore +2 this proposal from my end if the license headers are fixed.

Patch-set: 8
Label: Verified=0