Fix JcServiceStatusMXBean definition
The MBean definition was broken with out switch to Object keys
for job queue map. Unfortunately the implementation did not fail
when the platform server rejected the registration because of this
and the problem went undetected.

Fix the definition by explicitly dealing with Object->String
translation. If we encounter 'toString()' duplicates we end up
issuing a warning to fix the key class.

Refactor registration code to throw IllegalStateException when we
are rejected because the definition is broken.

Finally fix unit test suite to properly initialize JobCoordinator,
resulting in JMX being registered -- and thus fixing unregistration
warnings we have been getting.

JIRA: INFRAUTILS-74
Change-Id: I630e7526ff3271cf9bc32315ef740f7131d293ae
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>