Deprecate ClusterMemberInfo for removal
This facility is used only by opeflowplugin.git and it its
implementation has a strong binding to Akka without ever expressing that
anywhere in the build system.

So while 'infrautils must not depend on anything but odlparent', it
happily makes assumptions about cluster support being there, it being
based on Akka and Akka being configured to register with JMX.

Deprecate the interface along with its implementations for speedy
removal.

Change-Id: I7f3c1fa7ff11045f175a55c729dbf4f3da02f17d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>